+++ app/src/main/java/com/caffeine/capin/review/write/WriteReviewControllerImpl.kt
@@ -2,6 +2,7 @@ package com.caffeine.capin.review.write
 
 import com.caffeine.capin.network.BaseResponse
 import com.caffeine.capin.network.CapinApiService
+import io.reactivex.rxjava3.core.Completable
 import io.reactivex.rxjava3.core.Single
 import okhttp3.MultipartBody
 import okhttp3.RequestBody
@@ -10,6 +11,6 @@ import javax.inject.Inject
 class WriteReviewControllerImpl @Inject constructor(
     private val capinApiService: CapinApiService
 ): WriteReviewController {
-    override fun postReview(cafeId: String, review: MultipartBody.Part, imgs: List<MultipartBody.Part?>): Single<BaseResponse> =
+    override fun postReview(cafeId: String, review: MultipartBody.Part, imgs: List<MultipartBody.Part?>): Completable =
         capinApiService.postReview(cafeId, review,imgs)
 }
