+++ app/src/main/java/com/caffeine/capin/review/write/WriteReviewViewModel.kt
@@ -10,9 +10,11 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import com.caffeine.capin.PictureUriEntity
 import com.caffeine.capin.util.FormDataUtil
+import com.caffeine.capin.util.FormDataUtil.asMultipart
 import com.caffeine.capin.util.JsonStringParser
 import dagger.hilt.android.lifecycle.HiltViewModel
 import io.reactivex.rxjava3.android.schedulers.AndroidSchedulers
+import io.reactivex.rxjava3.core.Single
 import io.reactivex.rxjava3.schedulers.Schedulers
 import okhttp3.MediaType
 import okhttp3.MediaType.Companion.toMediaType
@@ -40,6 +42,10 @@ class WriteReviewViewModel @Inject constructor(
     val recommendation: LiveData<MutableList<Int?>>
         get() = _recommendation
 
+    private val _successPost = MutableLiveData<Boolean>()
+    val successPost: LiveData<Boolean>
+        get() = _successPost
+
     private val _checkedRecommend = MutableLiveData<Map<CompoundButton, Int>>()
     val checkedRecommend: LiveData<Map<CompoundButton, Int>>
         get() = _checkedRecommend
@@ -47,7 +53,6 @@ class WriteReviewViewModel @Inject constructor(
     fun changeCheckedRecommend(map: Map<CompoundButton, Int>) {
         _checkedRecommend.value = map
     }
-
 
     fun addImagesOfCafe(pictureUriEntity: PictureUriEntity) {
         cafePhotos.add(pictureUriEntity)
@@ -90,41 +95,20 @@ class WriteReviewViewModel @Inject constructor(
             requestMap.add(picture.uri?.asMultipart("imgs",contentResolver))
         }
 
-        //Todo: 플로우 연결 후 카페 아이디 하드코딩 수정정
+        //Todo: 플로우 연결 후 카페 아이디 하드코딩 수정
+
        writeReviewController.postReview(
-            "60e96789868b7d75f394b06f",
+            "60e96789868b7d75f394b06a",
             reviewJson,
             requestMap
         ).subscribeOn(Schedulers.io())
             .observeOn(AndroidSchedulers.mainThread())
             .subscribe({
-                Log.e("success", "fdfsad")
+                _successPost.postValue(true)
             }, {
                 it.printStackTrace()
-                Log.e("fail", "dfasd")
+                _successPost.postValue(false)
             })
     }
 
-
-    fun Uri.asMultipart(name: String, contentResolver: ContentResolver): MultipartBody.Part? {
-        return contentResolver.query(this, null, null, null, null)?.let {
-            if (it.moveToNext()) {
-                val displayName = it.getString(it.getColumnIndex(OpenableColumns.DISPLAY_NAME));
-                val requestBody = object : RequestBody() {
-                    override fun contentType(): MediaType? {
-                        return contentResolver.getType(this@asMultipart)?.toMediaType()
-                    }
-
-                    override fun writeTo(sink: BufferedSink) {
-                        sink.writeAll(contentResolver.openInputStream(this@asMultipart)?.source()!!)
-                    }
-                }
-                it.close()
-                MultipartBody.Part.createFormData(name, displayName, requestBody)
-            } else {
-                it.close()
-                null
-            }
-        }
-    }
 }
