+++ app/src/main/java/com/caffeine/capin/util/FormDataUtil.kt
@@ -1,8 +1,15 @@
 package com.caffeine.capin.util
 
+import android.content.ContentResolver
+import android.net.Uri
+import android.provider.OpenableColumns
+import okhttp3.MediaType
 import okhttp3.MediaType.Companion.toMediaType
 import okhttp3.MultipartBody
+import okhttp3.RequestBody
 import okhttp3.RequestBody.Companion.asRequestBody
+import okio.BufferedSink
+import okio.source
 import java.io.File
 
 object FormDataUtil {
@@ -10,12 +17,25 @@ object FormDataUtil {
         return MultipartBody.Part.createFormData(key, value.toString())
     }
 
-    fun getImageBody(key: String, file: File): MultipartBody.Part {
-        return MultipartBody.Part.createFormData(
-            name = key,
-            filename = file.name,
-            body = file.asRequestBody("multipart/form-data".toMediaType())
-        )
-    }
+    fun Uri.asMultipart(name: String, contentResolver: ContentResolver): MultipartBody.Part? {
+        return contentResolver.query(this, null, null, null, null)?.let {
+            if (it.moveToNext()) {
+                val displayName = it.getString(it.getColumnIndex(OpenableColumns.DISPLAY_NAME));
+                val requestBody = object : RequestBody() {
+                    override fun contentType(): MediaType? {
+                        return contentResolver.getType(this@asMultipart)?.toMediaType()
+                    }
 
+                    override fun writeTo(sink: BufferedSink) {
+                        sink.writeAll(contentResolver.openInputStream(this@asMultipart)?.source()!!)
+                    }
+                }
+                it.close()
+                MultipartBody.Part.createFormData(name, displayName, requestBody)
+            } else {
+                it.close()
+                null
+            }
+        }
+    }
 }
